try to fix CI
