+++ .nyc_output/processinfo/08eb685c-1776-4c17-85b4-314aa96435d6.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43759,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/4f9755e5931ce271f70f220646c92a1d9dd56baf","time":1588631091812,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/08eb685c-1776-4c17-85b4-314aa96435d6.json","externalId":"","uuid":"08eb685c-1776-4c17-85b4-314aa96435d6","files":[]}
