+++ .nyc_output/processinfo/1683ef69-d41c-4383-a422-659ecef74890.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43727,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g","--prefix","/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/10e7419e526681110dab6961d06313d6b2794027/.yo-repository","--loglevel","error"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/10e7419e526681110dab6961d06313d6b2794027","time":1588631084965,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/1683ef69-d41c-4383-a422-659ecef74890.json","externalId":"","uuid":"1683ef69-d41c-4383-a422-659ecef74890","files":[]}
