+++ .nyc_output/processinfo/44912daf-4a0e-4eaa-a2e3-eb169ab1378a.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43773,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/a546e04f4d5bec51fd3b0d5cddc915b618d949ef","time":1588631092846,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/44912daf-4a0e-4eaa-a2e3-eb169ab1378a.json","externalId":"","uuid":"44912daf-4a0e-4eaa-a2e3-eb169ab1378a","files":[]}
