+++ .nyc_output/processinfo/5b6afeff-d9fb-4e83-9ca9-3f6ea8552942.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43729,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/10e7419e526681110dab6961d06313d6b2794027","time":1588631085538,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/5b6afeff-d9fb-4e83-9ca9-3f6ea8552942.json","externalId":"","uuid":"5b6afeff-d9fb-4e83-9ca9-3f6ea8552942","files":[]}
