+++ .nyc_output/processinfo/5de2aa75-48d3-4d79-8c4e-9152ad828cce.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43755,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/da1b397d6365f53d98a59b711499ad271207010f","time":1588631090875,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/5de2aa75-48d3-4d79-8c4e-9152ad828cce.json","externalId":"","uuid":"5de2aa75-48d3-4d79-8c4e-9152ad828cce","files":[]}
