+++ .nyc_output/processinfo/8b6e67b4-e65f-4a8a-94ed-303a4d6c319b.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43761,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/9c213f180b9ff080866d909156f5d69893064ae3","time":1588631092170,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/8b6e67b4-e65f-4a8a-94ed-303a4d6c319b.json","externalId":"","uuid":"8b6e67b4-e65f-4a8a-94ed-303a4d6c319b","files":[]}
