+++ .nyc_output/processinfo/935535ef-47ba-4d15-b413-ce2c9bfeb2e1.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43750,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/8d398e973779f15f32cd151d0643bb2ac8ea521f","time":1588631089890,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/935535ef-47ba-4d15-b413-ce2c9bfeb2e1.json","externalId":"","uuid":"935535ef-47ba-4d15-b413-ce2c9bfeb2e1","files":[]}
