+++ .nyc_output/processinfo/954bf6cb-c6bf-45db-b965-f62ad768e5c9.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43752,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/00f020be06b33eb0b1e990ad24189b498f4983fe","time":1588631090389,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/954bf6cb-c6bf-45db-b965-f62ad768e5c9.json","externalId":"","uuid":"954bf6cb-c6bf-45db-b965-f62ad768e5c9","files":[]}
