+++ .nyc_output/processinfo/b6db1072-1786-437e-ac66-895339dfd0ac.json
@@ -1,0 +1,1 @@
+{"parent":null,"pid":43716,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/workspace/jhipster-quarkus/node_modules/.bin/mocha","--timeout","300000","--slow","0","test/client.spec.js","test/common.spec.js","test/entity-client.spec.js","test/entity-i18n.spec.js","test/entity-server.spec.js","test/languages.spec.js","test/server.spec.js","test/spring-controller.spec.js","test/spring-service.spec.js"],"execArgv":[],"cwd":"/Users/daniel/workspace/jhipster-quarkus","time":1588631081182,"ppid":43715,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/b6db1072-1786-437e-ac66-895339dfd0ac.json","externalId":"","uuid":"b6db1072-1786-437e-ac66-895339dfd0ac","files":["/Users/daniel/workspace/jhipster-quarkus/generators/client/index.js","/Users/daniel/workspace/jhipster-quarkus/generators/client/files.js","/Users/daniel/workspace/jhipster-quarkus/generators/common/index.js","/Users/daniel/workspace/jhipster-quarkus/generators/entity-client/index.js","/Users/daniel/workspace/jhipster-quarkus/generators/entity-i18n/index.js","/Users/daniel/workspace/jhipster-quarkus/generators/entity-server/index.js","/Users/daniel/workspace/jhipster-quarkus/generators/entity-server/files.js","/Users/daniel/workspace/jhipster-quarkus/generators/utils.js","/Users/daniel/workspace/jhipster-quarkus/generators/entity-server/prompts.js","/Users/daniel/workspace/jhipster-quarkus/generators/generator-quarkus-constants.js","/Users/daniel/workspace/jhipster-quarkus/generators/languages/index.js","/Users/daniel/workspace/jhipster-quarkus/generators/server/index.js","/Users/daniel/workspace/jhipster-quarkus/generators/server/prompts.js","/Users/daniel/workspace/jhipster-quarkus/generators/server/files.js","/Users/daniel/workspace/jhipster-quarkus/generators/spring-controller/index.js","/Users/daniel/workspace/jhipster-quarkus/generators/spring-service/index.js"]}
