+++ .nyc_output/processinfo/cc26f319-4d97-4366-afc6-49be5a2b6069.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43732,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/39246e015d1d5ceff4ba4812bc40cd3eaf525dc3","time":1588631086803,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/cc26f319-4d97-4366-afc6-49be5a2b6069.json","externalId":"","uuid":"cc26f319-4d97-4366-afc6-49be5a2b6069","files":[]}
