+++ .nyc_output/processinfo/cdb04c21-d466-4128-9b41-ece3945476ae.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43748,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/d7b6763d50d0bff2ae8c56117bf0f41e0c611f0d","time":1588631089442,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/cdb04c21-d466-4128-9b41-ece3945476ae.json","externalId":"","uuid":"cdb04c21-d466-4128-9b41-ece3945476ae","files":[]}
