+++ .nyc_output/processinfo/e6e5afce-01ee-4824-bb15-c98b1b5fbc4e.json
@@ -1,0 +1,1 @@
+{"parent":"b6db1072-1786-437e-ac66-895339dfd0ac","pid":43757,"argv":["/Users/daniel/.nvm/versions/node/v12.14.0/bin/node","/Users/daniel/.nvm/versions/node/v12.14.0/bin/npm","root","-g"],"execArgv":[],"cwd":"/private/var/folders/_y/4dzs9_nn3dggply5gbpwpmtm0000gn/T/e2fc27acb4a2b01462e28ef2f7c60f0dbbfdd0b0","time":1588631091350,"ppid":43716,"coverageFilename":"/Users/daniel/workspace/jhipster-quarkus/.nyc_output/e6e5afce-01ee-4824-bb15-c98b1b5fbc4e.json","externalId":"","uuid":"e6e5afce-01ee-4824-bb15-c98b1b5fbc4e","files":[]}
