+++ package.json
@@ -39,15 +39,18 @@
         "eslint-config-prettier": "4.3.0",
         "eslint-plugin-import": "2.17.3",
         "eslint-plugin-prettier": "3.1.0",
-        "fs-extra": "8.0.1",
         "husky": "^4.2.3",
-        "jsdoc": "3.6.3",
-        "mocha": "6.1.4",
         "pre-commit": "1.2.2",
         "prettier": "1.17.1",
         "prettier-plugin-java": "0.7.1",
+        "fs-extra": "9.0.0",
+        "jsdoc": "3.6.4",
+        "mocha": "7.1.2",
+        "nyc": "15.0.1",
+        "sinon": "9.0.2",
         "yeoman-assert": "3.1.1",
-        "yeoman-test": "2.2.0"
+        "yeoman-test": "2.4.1",
+        "sonarqube-scanner": "2.6.0"
     },
     "scripts": {
         "prettier:format": "prettier --write \"{,**/}*.{js,json,md,yml,java}\"",
@@ -57,8 +60,9 @@
         "ejs-lint": "npm run test:unit -- test/ejslint.js",
         "ejslint": "ejslint",
         "pretest": "npm run lint",
+        "sonar:scanner": "sonar-scanner",
         "test": "npm run test:unit -- test/*.spec.js",
-        "test:unit": "mocha --timeout 20000 --slow 0 --reporter spec",
+        "test:unit": "nyc --reporter=lcov mocha --timeout 300000 --slow 0",
         "jsdoc": "jsdoc --configure jsdoc-conf.json",
         "release": "git push && git push --tags && npm publish",
         "release:patch": "npm version patch -a -m \"Update to %s\" && npm run release",
