+++ test-integration/scripts/03-copy-entities.sh
@@ -13,7 +13,7 @@ mkdir -p "$JHI_FOLDER_APP"/.jhipster
 echo "===="
 echo "$JHI_SAMPLES"
 echo "===="
-ls .
+ls "$JHI_SAMPLES"/.jhipster/
 ls "$JHI_SAMPLES"/.jhipster/"$JHI_ENTITY"/*.json
 
 cp -f "$JHI_SAMPLES"/.jhipster/"$JHI_ENTITY"/*.json "$JHI_FOLDER_APP"/.jhipster/
